Add new validator and fixer for unused service groups

+++ palo_alto_firewall_analyzer/fixers/delete_unused_objects.py
@@ -15,20 +15,23 @@ def delete_unused_object(profilepackage, object_type, object_friendly_type, vali
 
     results_to_delete = validator_function(profilepackage)
 
-    print (f"Deleting {len(results_to_delete)} unused {object_friendly_type} objects now")
-    for badentry in results_to_delete:
-        device_group = badentry.device_group
-        object_name = badentry.data[0].get('name')
-        print(f"Deleting {object_name} from {device_group}")
-        try:
-            if device_group == 'shared':
-                pan_api.delete_object(panorama, version, api_key, object_type, object_name, 'shared')
-            else:
-                pan_api.delete_object(panorama, version, api_key, object_type, object_name, 'device-group', device_group)
-        except requests.HTTPError as err:
-            print(f"Error deleting {object_name} from {device_group}: {err.response.text}")
-    pan_api.validate_commit(panorama, api_key)
-    print (f"Deletion of {len(results_to_delete)} unused {object_friendly_type} objects complete. Please commit in the firewall.")
+    if results_to_delete:
+        print (f"Deleting {len(results_to_delete)} unused {object_friendly_type} objects now")
+        for badentry in results_to_delete:
+            device_group = badentry.device_group
+            object_name = badentry.data[0].get('name')
+            print(f"Deleting {object_name} from {device_group}")
+            try:
+                if device_group == 'shared':
+                    pan_api.delete_object(panorama, version, api_key, object_type, object_name, 'shared')
+                else:
+                    pan_api.delete_object(panorama, version, api_key, object_type, object_name, 'device-group', device_group)
+            except requests.HTTPError as err:
+                print(f"Error deleting {object_name} from {device_group}: {err.response.text}")
+        pan_api.validate_commit(panorama, api_key)
+        print (f"Deletion of {len(results_to_delete)} unused {object_friendly_type} objects complete. Please commit in the firewall.")
+    else:
+        print (f"There were no {object_friendly_type} to delete")
     return results_to_delete
 
 @register_policy_fixer("DeleteUnusedAddresses", "Delete Address objects that aren't in use")
@@ -52,10 +55,18 @@ def delete_unused_services(profilepackage):
     _, _, validator_function = get_policy_validators()['UnusedServices']
     return delete_unused_object(profilepackage, object_type, object_friendly_type, validator_function)
 
-@register_policy_fixer("DeleteUnusedOjects", "Convenience wrapper that calls DeleteUnusedAddressGroups, DeleteUnusedAddresses, and DeleteUnusedServices")
-def delete_unused_services(profilepackage):
+@register_policy_fixer("DeleteUnusedServiceGroups", "Delete Service Group objects that aren't in use")
+def delete_unused_servicegroups(profilepackage):
+    object_type = "ServiceGroups"
+    object_friendly_type = "Service Groups"
+    _, _, validator_function = get_policy_validators()['UnusedServiceGroups']
+    return delete_unused_object(profilepackage, object_type, object_friendly_type, validator_function)
+
+@register_policy_fixer("DeleteUnusedOjects", "Convenience wrapper that calls DeleteUnusedAddressGroups, DeleteUnusedAddresses, DeleteUnusedServiceGroups, and DeleteUnusedServices")
+def delete_unused_objects(profilepackage):
     deleted_objects = []
     deleted_objects += delete_unused_addressgroups(profilepackage)
     deleted_objects += delete_unused_addresses(profilepackage)
+    deleted_objects += delete_unused_servicegroups(profilepackage)
     deleted_objects += delete_unused_services(profilepackage)
     return deleted_objects
